Reorganize baseline code, and use that codebase to generate submissions and evaluations.
