+++ object-detection/baseline/coco.py
