+++ object-detection/baseline/group_by_aspect_ratio.py
